Enable stable flutter (#14)
* Bump dart version and enable analyzer

Dart 2.19.0 fixes an issue where named pipes read as stream using a File
would end the stream too early.

* Run CI once a week, enable Flutter stable channel

* Work around pre-release versions not satisfying dart SDK constraint

* Remove deprecated Flutter dev branch

From the wiki:

There used to be a dev branch as well (and you may still see it in our
tooling). We no longer update this branch and our tooling will be
updated in due course to no longer list it (if it hasn't already).